django_telegram_news/pyproject.toml
[tool.poetry]
name = "django-telegram"
version = "0.1.0"
description = ""
authors = ["Qiyas <107604130+qiyascc@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.1"
DateTime = "^5.4"
Pyrogram = "^2.0.106"
TgCrypto = "^1.2.5"
djangorestframework = "^3.14.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
